tax to int

+++ app.py
@@ -52,7 +52,7 @@ def handle_pattern(text):
 		keyword = text.split(' ')[0]
 		value = text.split(' ')[1]
 		ret = {
-			'tax': str(float(value)*110/100),
+			'tax': str(int(float(value)*110/100)),
 			'serv': str(int(float(value)*105*110/10000)),
 			'service': str(int(float(value)*105*110/10000))
 		}.get(keyword, '')
